delete button no longer errors and deletes correctly from db

+++ server/api/queue/job/[jobId].delete.ts
@@ -2,7 +2,7 @@ import { eq } from "drizzle-orm";
 import * as z from "zod";
 
 import { mysqlDb } from "../../../db/mysql";
-import { jobTable } from "../../../db/mysql/schema";
+import { cohortResultsTable, jobTable, queryResultSetTable, queryResultTable } from "../../../db/mysql/schema";
 
 const paramSchema = z.object({
   jobId: z.string()
@@ -17,8 +17,28 @@ export default defineEventHandler(async event => {
     .where(eq(jobTable.id, Number(jobId)));
   const item = items[0];
   if (item) {
-    await mysqlDb.delete(jobTable).where(eq(jobTable.id, item.id));
+    const resultSetIds = await mysqlDb.select().from(queryResultSetTable).where(eq(queryResultSetTable.jobId, item.id));
+    const resultSetId = resultSetIds[0];
+    if (resultSetId) {
+      const resultIds = await mysqlDb.select().from(queryResultTable).where(eq(queryResultTable.queryResultSetId, resultSetId.id));
+      const resultId = resultIds[0];
+      if (resultId) {
+        try {
+          await mysqlDb.delete(cohortResultsTable).where(eq(cohortResultsTable.queryResultId, resultId.id));
+          await mysqlDb.delete(queryResultTable).where(eq(queryResultTable.id, resultId.id));
+          await mysqlDb.delete(queryResultSetTable).where(eq(queryResultSetTable.jobId, item.id));
+          await mysqlDb.delete(jobTable).where(eq(jobTable.id, item.id));
+        } catch (err) {
+          console.error("Error deleting job id:", item.id, err);
+          throw err;
+        }
+      } else {
+        createError("Query result data not found for job id: " + jobId);
+      }
+    } else {
+      createError("Query result set data not found for job id: " + jobId);
+    }
   } else {
-    createError("Query queue item not found for id: " + jobId);
+    createError("Query queue item not found for job id: " + jobId);
   }
 });
